alice_control_panel/requirements.txt
aiohttp
fastapi
faster-whisper
google-auth
numpy
onnxruntime
openai
psutil
PyYAML
requests
uvicorn[standard]

